+++ CHANGES.rst
@@ -10,6 +10,10 @@ Changelog
 3.0.3 (2021-01-27)
 ------------------
 
+- Ensure plone.app.discussion comments with acquired keywords are
+  consistently reindexed.
+  [alecpm]
+
 - Fix toggle to work with Chameleon.
   Fixes `issue #33 <https://github.com/collective/Products.PloneKeywordManager/issues/33>`_.
   [petschki]